+++ local-addons/morons/models/__init__.py
@@ -2,4 +2,6 @@
 
 from . import project
 from . import contributor
+from . import sale_model
 from . import invoice
+from . import clientInvoice
